+++ packages/core/src/components/TextField/TextField.tsx
@@ -20,35 +20,27 @@ export const TextField: SFC<Props> & WithStyle = React.memo(
                 required,
                 disabled,
                 placeholder,
+                validator,
                 ...restProps
             } = props,
-            inputId = id || 'medly-textField',
-            isLabelPresent = !!label,
-            isPrefixPresent = !!Prefix,
-            isSuffixPresent = !!Suffix,
-            isErrorPresent = !!errorText || !!builtInErrorMessage;
+            inputId = useMemo(() => id || 'medly-textField', [id]),
+            isLabelPresent = useMemo(() => !!label, [label]),
+            isPrefixPresent = useMemo(() => !!Prefix, [Prefix]),
+            isSuffixPresent = useMemo(() => !!Suffix, [Suffix]),
+            isErrorPresent = useMemo(() => !!errorText || !!builtInErrorMessage, [errorText, builtInErrorMessage]);
+
+        const validate = (event: React.FormEvent<HTMLInputElement>, eventFunc: (e: React.FormEvent<HTMLInputElement>) => void) => {
+            event.preventDefault();
+            const element = event.target as HTMLInputElement,
+                message = (validator && validator(element.value)) || element.validationMessage;
+            setErrorMessage(message);
+            eventFunc && eventFunc(event);
+        };
 
         const stopPropagation = useCallback((event: React.MouseEvent) => event.stopPropagation(), []),
-            handleWrapperClick = useCallback(() => {
-                !disabled && inputRef.current.focus();
-            }, [inputRef, disabled]),
-            onBlur = useCallback((event: React.FocusEvent<HTMLInputElement>) => {
-                setErrorMessage((event.target as HTMLInputElement).validationMessage);
-                props.onBlur && props.onBlur(event);
-            }, []),
-            onInvalid = useCallback((event: React.FormEvent<HTMLInputElement>) => {
-                event.preventDefault();
-                setErrorMessage((event.target as HTMLInputElement).validationMessage);
-                props.onInvalid && props.onInvalid(event);
-            }, []),
-            builtInFormValidators = useMemo(
-                () => ({
-                    onBlur,
-                    onInvalid,
-                    errorText: errorText || builtInErrorMessage
-                }),
-                [onBlur, onInvalid, errorText, builtInErrorMessage]
-            );
+            handleWrapperClick = useCallback(() => !disabled && inputRef.current.focus(), [inputRef, disabled]),
+            onBlur = useCallback((event: React.FocusEvent<HTMLInputElement>) => validate(event, props.onBlur), []),
+            onInvalid = useCallback((event: React.FormEvent<HTMLInputElement>) => validate(event, props.onInvalid), []);
 
         return (
             <Styled.OuterWrapper fullWidth={fullWidth} minWidth={minWidth}>
@@ -75,7 +67,8 @@ export const TextField: SFC<Props> & WithStyle = React.memo(
                             isPrefixPresent={isPrefixPresent}
                             isSuffixPresent={isSuffixPresent}
                             isLabelPresent={isLabelPresent}
-                            {...{ ...restProps, ...(props.withBuiltInValidation && builtInFormValidators) }}
+                            errorText={errorText || builtInErrorMessage}
+                            {...{ ...restProps, onBlur, onInvalid }}
                         />
                         <Styled.Label htmlFor={`${inputId}-input`} required={required} variant={props.variant}>
                             {label}
@@ -101,7 +94,6 @@ TextField.Style = Styled.OuterWrapper;
 TextField.defaultProps = {
     type: 'text',
     variant: 'filled',
-    withBuiltInValidation: false,
     fullWidth: false,
     disabled: false,
     required: false,