+++ packages/core/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-beta.34](https://github.com/medly/medly-components/compare/@medly-components/core@1.0.0-beta.33...@medly-components/core@1.0.0-beta.34) (2020-06-01)
+
+
+### Bug Fixes
+
+* remove few deprecated storybook functionality ([b1a12bd](https://github.com/medly/medly-components/commit/b1a12bd347550b56e8b7985cbd0013a56108c5dd))
+
+
+### Features
+
+* add validator function in TextField component ([eaf15f6](https://github.com/medly/medly-components/commit/eaf15f63529b8d7d484ddfd43390a2aa273e1700))
+* add validator function in TextField component ([dc6e83c](https://github.com/medly/medly-components/commit/dc6e83ccec8ae403ae2892d85bf9737fcb45e4c9))
+
+
+
+
+
 # [1.0.0-beta.33](https://github.com/medly/medly-components/compare/@medly-components/core@1.0.0-beta.32...@medly-components/core@1.0.0-beta.33) (2020-05-31)